Create sudoku_and_web.py
This is a little project I did on a Saturday afternoon after I read, that a Sudoku solver is a good example for a quick back tracking algorithm. When the algorithm was ready, I found it a bit tedious to look up and enter new boards manually all the time, so I looked up a free sudoku puzzle website and used selenium to read in the puzzle and put in the solution.